# src/docs/swagger/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
  models.AuthToken:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
    type: object
host: '{{.Host}}'
info:
  contact: {}
  description: Zeroslope microservice architecture written in go
  license: {}
  title: ZeroSlope API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Generates a JWT Token for use with Authorized endpoints
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthToken'
            type: object
        "401": {}
      summary: Login generates a JWT token
      tags:
      - Authentication
  /health:
    get:
      consumes:
      - application/json
      description: Returns a message if the service is working correctly
      produces:
      - application/json
      responses:
        "200":
          description: The service is working.
          schema:
            type: string
      summary: Server Health
      tags:
      - Health
  /samples:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: Delete removes a sample by id
      tags:
      - Sample
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      tags:
      - Sample
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: Create stores a sample model
      tags:
      - Sample
    put:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: Update updates a sample model
      tags:
      - Sample
  /samples/:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: Read returns a single sample by id
      tags:
      - Sample
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"